+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Config API url via the config file `~/.sosrc`
+
 ## [1.0.4] - 2022-07-18
 ### Fixed
 - Uploading single file applet with front-applet version